Tighten worktree cleanup implementation layout
